matrikkel/matrikkel_search/requirements.txt
--extra-index-url https://europe-west3-python.pkg.dev/innbyggerkontakt-dev/ik-python-repo/simple
--extra-index-url https://europe-west3-python.pkg.dev/innbyggerkontakt/ik-python-repo/simple

blinker==1.6.2 ; python_version >= "3.10" and python_version < "4"
cachetools==5.3.1 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.5.7 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.10" and python_version < "4"
cloud-event-model==0.1.5 ; python_version >= "3.10" and python_version < "4.0"
cloudevents==1.9.0 ; python_version >= "3.10" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4" and platform_system == "Windows"
deprecated==1.2.14 ; python_version >= "3.10" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.10" and python_version < "4"
flask==2.3.2 ; python_version >= "3.10" and python_version < "4"
functions-framework==3.4.0 ; python_version >= "3.10" and python_version < "4"
google-api-core==2.11.1 ; python_version >= "3.10" and python_version < "4.0"
google-api-core[grpc]==2.11.1 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.21.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-core==2.3.2 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-pubsub==2.17.1 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-secret-manager==2.16.1 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-storage==2.10.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-trace==1.11.1 ; python_version >= "3.10" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
google-events==0.2.1 ; python_version >= "3.10" and python_version < "4.0"
google-resumable-media==2.5.0 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.59.1 ; python_version < "4.0" and python_version >= "3.10"
googleapis-common-protos[grpc]==1.59.1 ; python_version >= "3.10" and python_version < "4.0"
grpc-google-iam-v1==0.12.6 ; python_version >= "3.10" and python_version < "4.0"
grpcio==1.56.0 ; python_version < "4.0" and python_version >= "3.10"
grpcio-status==1.56.0 ; python_version < "4.0" and python_version >= "3.10"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "4" and platform_system != "Windows"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.10" and python_version < "4"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4"
markupsafe==2.1.3 ; python_version >= "3.10" and python_version < "4"
msgspec==0.9.1 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-api==1.13.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-exporter-gcp-trace==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation==0.34b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-flask==0.34b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-requests==0.34b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-wsgi==0.34b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-propagator-gcp==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-sdk==1.13.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-semantic-conventions==0.34b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-util-http==0.34b0 ; python_version >= "3.10" and python_version < "4.0"
otel-setup==0.2.3 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.10" and python_version < "4"
proto-plus==1.22.3 ; python_version >= "3.10" and python_version < "4.0"
protobuf==4.23.3 ; python_version < "4.0" and python_version >= "3.10"
ps-message-matrikkel-owner-search==0.1.3 ; python_version >= "3.10" and python_version < "4.0"
ps-message-matrikkel-search-event==0.1.3 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.9 ; python_version >= "3.10" and python_version < "4.0"
pyproj==3.6.0 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
setuptools==68.0.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
suds==1.1.2 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.7.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.16 ; python_version >= "3.10" and python_version < "4.0"
watchdog==3.0.0 ; python_version >= "3.10" and python_version < "4"
werkzeug==2.3.6 ; python_version >= "3.10" and python_version < "4"
wrapt==1.15.0 ; python_version >= "3.10" and python_version < "4.0"
